refactor: unify opamp client builder
