--- super-agent/src/sub_agent/opamp/mod.rs
@@ -1,3 +1,0 @@
-pub mod client_builder;
-// pub mod common;
-pub mod remote_config_publisher;
