+++ super-agent/src/super_agent/mod.rs
@@ -2,7 +2,6 @@ pub mod config;
 pub mod defaults;
 pub mod error;
 pub(super) mod event_handler;
-pub mod opamp;
 pub mod store;
 
 pub use super_agent::*;
