--- super-agent/src/super_agent/opamp/mod.rs
@@ -1,2 +1,0 @@
-pub mod client_builder;
-pub mod remote_config_publisher;
